Safari-609.1.11.1/LayoutTests/js/read-modify-eval-expected.txt

Tests whether eval() works inside statements that read and modify a value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS multTest(); is true
PASS divTest(); is true
PASS addTest(); is true
PASS subTest(); is true
PASS lshiftTest(); is true
PASS rshiftTest(); is true
PASS urshiftTest(); is true
PASS andTest(); is true
PASS xorTest(); is true
PASS orTest(); is true
PASS modTest(); is true
PASS preIncTest(); is true
PASS preDecTest(); is true
PASS postIncTest(); is true
PASS postDecTest(); is true
PASS primitiveThisTest.call(1); is true
PASS strictThisTest.call(1); threw exception TypeError: Attempted to assign to readonly property..
PASS successfullyParsed is true

TEST COMPLETE
